@@ -67,7 +67,7 @@ class GoodInfluxUdp extends Stream.Writable {
             lineData = lineData.concat(lines);
         });
 
-        const payload = new Buffer(lineData.join('\n'));
+        const payload = Buffer.from(lineData.join('\n'));
 
         // Prevent this from user tampering
         this._udpClient.send(payload, 0, payload.length, this._endpoint.port, this._endpoint.hostname, callback);
