@@ -79,7 +79,7 @@ class GoodInfluxHttp extends Stream.Writable {
         Hoek.merge(wreckOptions, this._settings.wreck, { nullOverride: false });
 
         // Prevent this from user tampering
-        if (wreckOptions.headers){
+        if (wreckOptions.headers) {
             wreckOptions.headers['content-type'] = 'text/plain';
         }
         else {
