@@ -1,7 +1,7 @@
 'use strict';
 
-const Code = require('code');
-const Lab = require('lab');
+const Code = require('@hapi/code');
+const Lab = require('@hapi/lab');
 const lab = exports.lab = Lab.script();
 
 const describe = lab.describe;
@@ -11,25 +11,23 @@ const expect = Code.expect;
 const Helpers = require('../lib/helpers');
 
 describe('measurementPrefix', () => {
-    it('should preserve raw strings', (done) => {
+    it('should preserve raw strings', () => {
         const config = {
             prefix: 'my-awesome-service-'
         };
         const prefix = Helpers.measurementPrefix(config);
 
         expect(prefix).to.equal('my-awesome-service-');
-        done();
     });
-    it('should join string arrays with a / by default', (done) => {
+    it('should join string arrays with a / by default', () => {
         const config = {
             prefix: ['my', 'awesome', 'service']
         };
         const prefix = Helpers.measurementPrefix(config);
 
         expect(prefix).to.equal('my/awesome/service/');
-        done();
     });
-    it('should join string arrays with a specified delimiter', (done) => {
+    it('should join string arrays with a specified delimiter', () => {
         const config = {
             prefix: ['my', 'awesome', 'service'],
             prefixDelimiter: '/'
@@ -37,16 +35,13 @@ describe('measurementPrefix', () => {
         const prefix = Helpers.measurementPrefix(config);
 
         expect(prefix).to.equal('my/awesome/service/');
-        done();
     });
-    it('should return an empty string, given an empty config', (done) => {
+    it('should return an empty string, given an empty config', () => {
         const prefix = Helpers.measurementPrefix({});
         expect(prefix).to.equal('');
-        done();
     });
-    it('should return an empty string, given no config', (done) => {
+    it('should return an empty string, given no config', () => {
         const prefix = Helpers.measurementPrefix();
         expect(prefix).to.equal('');
-        done();
     });
 });
