@@ -3,8 +3,8 @@
 const LineProtocol = require('../lib/line-protocol');
 const Schemas = require('../schemas');
 
-const Code = require('code');
-const Lab = require('lab');
+const Code = require('@hapi/code');
+const Lab = require('@hapi/lab');
 const lab = exports.lab = Lab.script();
 
 const describe = lab.describe;
@@ -12,7 +12,7 @@ const it = lab.it;
 const expect = Code.expect;
 
 describe('log', () => {
-    it('Event log is formatted as expected', (done) => {
+    it('Event log is formatted as expected', () => {
         const testEvent = {
             event: 'log',
             host: 'mytesthost',
@@ -24,12 +24,11 @@ describe('log', () => {
         const formattedLogEvent = LineProtocol.format(testEvent, {}, Schemas);
         const expectedLogEvent = ['log,host=mytesthost,pid=1234 data="Things are good",tags="info,request" 1485996802647000000'];
         expect(formattedLogEvent).to.equal(expectedLogEvent);
-        done();
     });
 });
 
 describe('request', () => {
-    it('Event request is formatted as expected', (done) => {
+    it('Event request is formatted as expected', () => {
         const testEvent = {
             event: 'request',
             host: 'mytesthost',
@@ -44,13 +43,12 @@ describe('request', () => {
         const formattedLogEvent = LineProtocol.format(testEvent, {}, Schemas);
         const expectedLogEvent = ['request,host=mytesthost,pid=1234 data="userid=001",id="4321",method="POST",path="/hello",tags="request,blahblahblah" 1485996802647000000'];
         expect(formattedLogEvent).to.equal(expectedLogEvent);
-        done();
     });
 });
 
 
 describe('response', () => {
-    it('Event response is formatted as expected', (done) => {
+    it('Event response is formatted as expected', () => {
         const testEvent = {
             event: 'response',
             host: 'mytesthost',
@@ -76,12 +74,11 @@ describe('response', () => {
         const formattedLogEvent = LineProtocol.format(testEvent, {}, Schemas);
         const expectedLogEvent = ['response,host=mytesthost,pid=1234 httpVersion="1.1",id="1234",instance="mytesthost",labels="label001",method="GET",path="/hello",query="k1=v1&k2=v2",referer="referer",remoteAddress="127.0.0.1",responseTime=500i,statusCode=200i,userAgent="Chrome" 1485996802647000000'];
         expect(formattedLogEvent).to.equal(expectedLogEvent);
-        done();
     });
 });
 
 describe('error', () => {
-    it('Event error is formatted as expected', (done) => {
+    it('Event error is formatted as expected', () => {
         const testEvent = {
             event: 'error',
             host: 'mytesthost',
@@ -100,19 +97,17 @@ describe('error', () => {
         const formattedLogEvent = LineProtocol.format(testEvent, {}, Schemas);
         const expectedLogEvent = ['error,host=mytesthost,pid=1234 error.name="error1",error.message="this is an error msg",error.stack="stackoverflow",id="1234",url="/hello",method="GET",tags="error,crash" 1485996802647000000'];
         expect(formattedLogEvent).to.equal(expectedLogEvent);
-        done();
     });
 });
 
 describe('Unrecognized event', () => {
-    it('LineProtocol simply returns', (done) => {
+    it('LineProtocol simply returns', () => {
         const testEvent = {
             event: 'hello',
             host: 'mytesthost',
             timestamp: 1485996802647
         };
         expect(LineProtocol.format(testEvent, {}, Schemas)).to.equal([]);
-        done();
     });
 });
 
@@ -127,28 +122,26 @@ describe('Measurement prefixes', () => {
         pid: 1234
     };
 
-    it('Are applied correctly', (done) => {
+    it('Are applied correctly', () => {
         const configs = {
             prefix: ['my', 'awesome', 'service']
         };
         const eventWithPrefix = ['my/awesome/service/log,host=mytesthost,pid=1234 data="Things are good",tags="info,request" 1485996802647000000'];
         expect(LineProtocol.format(testEvent, configs, Schemas)).to.equal(eventWithPrefix);
-        done();
     });
 
-    it('Special characters are escaped', (done) => {
+    it('Special characters are escaped', () => {
         const configs = {
             prefix: ['my', ' ', 'awesome', ',', 'service', ' '],
             prefixDelimiter: ''
         };
         const eventWithSpecialCharsInPrefix = ['my\\ awesome\\,service\\ log,host=mytesthost,pid=1234 data="Things are good",tags="info,request" 1485996802647000000'];
         expect(LineProtocol.format(testEvent, configs, Schemas)).to.equal(eventWithSpecialCharsInPrefix);
-        done();
     });
 });
 
 describe('configs', () => {
-    it('Event log is formatted as expected with default tags', (done) => {
+    it('Event log is formatted as expected with default tags', () => {
         const testEvent = {
             event: 'log',
             host: 'mytesthost',
@@ -164,10 +157,9 @@ describe('configs', () => {
         }, Schemas);
         const expectedLogEvent = ['log,test=test,host=mytesthost,pid=1234 data="Things are good",tags="info,request" 1485996802647000000'];
         expect(formattedLogEvent).to.equal(expectedLogEvent);
-        done();
     });
 
-    it('Event log is formatted as expected with default fields', (done) => {
+    it('Event log is formatted as expected with default fields', () => {
         const testEvent = {
             event: 'log',
             host: 'mytesthost',
@@ -183,10 +175,9 @@ describe('configs', () => {
         }, Schemas);
         const expectedLogEvent = ['log,host=mytesthost,pid=1234 test="test",data="Things are good",tags="info,request" 1485996802647000000'];
         expect(formattedLogEvent).to.equal(expectedLogEvent);
-        done();
     });
 
-    it('Event log is formatted as expected with event name redirection', (done) => {
+    it('Event log is formatted as expected with event name redirection', () => {
         const testEvent = {
             event: 'log',
             host: 'mytesthost',
@@ -206,10 +197,9 @@ describe('configs', () => {
         });
         const expectedLogEvent = ['test,host=mytesthost,pid=1234 data="Things are good",tags="info,request" 1485996802647000000'];
         expect(formattedLogEvent).to.equal(expectedLogEvent);
-        done();
     });
 
-    it('Event log is formatted as expected with event name function redirection', (done) => {
+    it('Event log is formatted as expected with event name function redirection', () => {
         const testEvent = {
             event: 'log',
             host: 'mytesthost',
@@ -229,6 +219,5 @@ describe('configs', () => {
         });
         const expectedLogEvent = ['test,host=mytesthost,pid=1234 data="Things are good",tags="info,request" 1485996802647000000'];
         expect(formattedLogEvent).to.equal(expectedLogEvent);
-        done();
     });
 });
